api/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
ansicon==1.89.0
anyio==3.6.1
APScheduler==3.9.1.post1
argcomplete==2.0.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.2.0
autobahn==22.7.1
Automat==22.10.0
autopep8==1.7.0
base58==2.1.1
bitarray==2.6.1
black==22.6.0
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
constantly==15.1.0
cryptography==38.0.3
cytoolz==0.12.1
daphne==4.0.0
Deprecated==1.2.13
discord.py==1.7.3
distlib==0.3.4
dj-database-url==0.5.0
Django==4.1.4
django-apscheduler==0.6.2
django-channels==0.7.0
django-cors-headers==3.13.0
django-filter==22.1
django-heroku==0.3.1
django-rest-framework==0.1.0
djangochannelsrestframework==1.1.0
djangorestframework==3.14.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.1
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
filelock==3.6.0
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.12.0
hexbytes==0.3.0
httpcore==0.15.0
httpx==0.23.0
hyperlink==21.0.0
idna==3.4
imutils==0.5.4
incremental==22.10.0
ipfshttpclient==0.8.0a2
jsonschema==4.17.3
lru-dict==1.1.8
Markdown==3.4.1
msgpack==1.0.4
msgpack-python==0.5.6
multiaddr==0.0.9
multidict==6.0.4
mypy-extensions==0.4.3
netaddr==0.8.0
numpy==1.23.2
nvidia-pyindex==1.0.9
oauthlib==3.2.2
opencv-python==4.6.0.66
packaging==21.3
parsimonious==0.8.1
pathspec==0.9.0
Pillow==9.0.1
pipx==1.1.0
platformdirs==2.5.1
protobuf==3.19.5
psycopg2==2.9.5
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
pycryptodome==3.16.0
pydantic==1.9.1
pyfakewebcam==0.1.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pyvirtualcam==0.9.1
pywin32==305
redis==4.4.0
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rlp==2.0.1
service-identity==21.1.0
shroomdk==1.0.2
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
Twisted==22.10.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.11
userpath==1.8.0
varint==1.0.2
virtualenv==20.13.2
virtualenvwrapper-win==1.2.7
wcwidth==0.2.5
web3==5.31.3
websockets==9.1
whitenoise==6.2.0
wrapt==1.14.1
yarl==1.8.2
zope.interface==5.5.2
